design: 헤더 및 배너 전체 가로길이 차지하도록 레이아웃 조정 #139
